Split binder tracing tests into separate test projects
